FIXED: Extended query solutions from `int` to `int64_t`
Otherwise, PL_new_solution() is limited to 2G solutions.

+++ src/pl-incl.h
@@ -1853,7 +1853,7 @@ struct queryFrame
   unsigned int	flags;
   debug_type	debugSave;		/* saved debugstatus.debugging */
   unsigned int	flags_saved;		/* Saved boolean Prolog flags */
-  int		solutions;		/* # of solutions produced */
+  int64_t	solutions;		/* # of solutions produced */
   Word	       *aSave;			/* saved argument-stack */
   Choice	saved_bfr;		/* Saved choice-point */
   LocalFrame	saved_ltop;		/* Saved lTop */
